Move GET /api/stats into backend/src/routes/stats.ts
Verbatim relocation, no logic changes. Part of #116.
